@@ -8,9 +8,14 @@ Bundler.require(*Rails.groups)
 
 module Starlistz
   class Application < Rails::Application
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.1
+
     # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
+    
     config.i18n.default_locale = :ja
     config.time_zone = 'Asia/Tokyo' #日本時間に変更
     config.active_record.default_timezone = :local #DB に書かれる時刻を日本語に
